feat: import and export commitments working in a simple way
