+++ wallet/src/components/Transactions/index.scss
@@ -15,7 +15,6 @@
   }
 
   &__right {
-    
     width: 85%;
   }
 }
@@ -119,31 +118,31 @@
         width: 22px;
         margin-top: -2px;
         margin-right: 2px;
-      }      
+      }
     }
 
     .fastWithdraw {
-        color: $light-primary-500;        
-        outline: none;
-        margin: 6px 0;          
-        border-radius: 38px;        
-        cursor: pointer;
-  
-        @media #{$lg-devices} {
-          width: fit-content;
-        }          
-  
-        .infoIcon {
-          height: 22px;
-          width: 22px;
-          margin-top: -2px;
-          margin-right: 2px;
-        }
-  
-        // &fastWithdraw {
-        font-weight: bold;
-        text-decoration: underline;        
+      color: $light-primary-500;
+      outline: none;
+      margin: 6px 0;
+      border-radius: 38px;
+      cursor: pointer;
+
+      @media #{$lg-devices} {
+        width: fit-content;
       }
+
+      .infoIcon {
+        height: 22px;
+        width: 22px;
+        margin-top: -2px;
+        margin-right: 2px;
+      }
+
+      // &fastWithdraw {
+      font-weight: bold;
+      text-decoration: underline;
+    }
   }
 }
 
@@ -201,28 +200,28 @@
 }
 
 .transactionTabs {
-  margin-right:  32px;
-  padding:  0;
-  border-radius:  unset;
-  font-weight:  800;
-  font-size:  16px;
-  line-height:  24px;
-  height:  44px;
-  background:  unset;
+  margin-right: 32px;
+  padding: 0;
+  border-radius: unset;
+  font-weight: 800;
+  font-size: 16px;
+  line-height: 24px;
+  height: 44px;
+  background: unset;
   color: #7b3fe4;
-  border-bottom:  2px solid #7b3fe4;
-  display:  flex;
+  border-bottom: 2px solid #7b3fe4;
+  display: flex;
 }
 
 .transactionRow {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: spaceBetween;
-    border: 1px solid #dddfe0;
-    box-sizing: borderBox;
-    border-radius: 12px;
-    margin: 0 0 8px;
-    cursor: pointer;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: spaceBetween;
+  border: 1px solid #dddfe0;
+  box-sizing: borderBox;
+  border-radius: 12px;
+  margin: 0 0 8px;
+  cursor: pointer;
 }
 
 .transactionDetails {
@@ -245,6 +244,38 @@
   box-sizing: border-box;
   box-shadow: 0 2px 2px -2px rgb(0 0 0 / 8%), inset 0 1px 0.5px hsl(0deg 0% 100% / 16%);
   border-radius: 8px;
+}
+
+.wrapperTabList {
+  display: flex;
+  justify-content: space-between;
+}
+
+.exportTransactionsButton {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  padding: 15px 20px;
+  margin: 12px 35px 0;
+
+  height: 38px;
+  left: 432px;
+  top: 694px;
+
+  /* light/primary-500 */
+
+  background: #7b3fe4;
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  box-sizing: border-box;
+  /* Button/Primary */
+  border-radius: 12px;
+
+  color: #fff;
+
+  &:hover {
+    background: #7635e6;
+  }
 }
 
 .tab-list {
